+++ docs/CHANGELOG.md
@@ -21,6 +21,7 @@
 - added new Lua catalogs for Lara states, Lara anims, music, and samples
 - added support for cross-fades to the title screen
 - added visual previews of bar colors (Graphic Options → Bars)
+- added the ability to change PS1 bar colors
 - added endless sprint (available previously via the `/restless` command) to the UI settings (Gameplay options → Mods → Endless sprint)
 - added endless flare time cheat (Gameplay options → Mods → Endless flare time)
 - added `O_VULTURE` for custom levels